App/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.6
prometheus-flask-exporter==0.20.1
Werkzeug==2.2.3
gunicorn==20.1.0
python-dotenv==1.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
pytest==7.3.1
pytest-flask==1.2.0
coverage==7.2.5
